+++ DawnLib/src/API/MapObjects/.MapObjectRegistrationHandler.cs
@@ -266,7 +266,7 @@ static class MapObjectRegistrationHandler
 
     internal static void UpdateInsideMapObjectSpawnWeightsOnLevel(SelectableLevel level)
     {
-        if (!LethalContent.MapObjects.IsFrozen || StartOfRound.Instance == null || (WeatherRegistryCompat.Enabled && WeatherRegistryCompat.IsWeatherManagerReady()))
+        if (!LethalContent.MapObjects.IsFrozen || StartOfRound.Instance == null || (WeatherRegistryCompat.Enabled && !WeatherRegistryCompat.IsWeatherManagerReady()))
             return;
 
         foreach (DawnMapObjectInfo mapObjectInfo in LethalContent.MapObjects.Values)