+++ DawnLib/src/API/Tags/SimpleAutoTagger.cs
@@ -3,7 +3,6 @@
 namespace Dawn;
 public class SimpleAutoTagger<T>(NamespacedKey tag, Func<T, bool> predicate) : IAutoTagger<T> where T : DawnBaseInfo<T>
 {
-
     public NamespacedKey Tag => tag;
     public bool ShouldApply(T info)
     {