+++ Thunderstore/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Added more config error checking to the Definitions as this mod uses a different format compared to other mods.
 - Added the tag `custom` to the mod as `lethal_company:custom` or just `custom` should work.
 - Entity, Enemy and Item content definitions have defaults weights of 0 on every vanilla moon, multipliers of 1 on every vanilla weather and default weight of 0 on every vanilla interior.
+- Fixed `SpawnWeightsPreset` entirely not working causing nothing to have any weights.
+- Fixed Weight Updater depending on weather never running thus nothing causing everything to not have any weights.
 
 ## v0.2.5